adds permission-debied cmd
